refactor: 메인 메소드 최소화
